+++ Help.ps1
@@ -171,3 +171,14 @@ $Environment = @'
 		}
 	)
 }
+
+### Reset-Rest
+@{
+	command = 'Reset-Rest'
+	synopsis = 'Resets the module state.'
+	description = @'
+	It tells to reset the module state without reloading it or restarting the
+	session. This includes cached results of named requests. After Reset-Rest
+	named requests are invoked once again.
+'@
+}