+++ src/PSRest/Commands/ResetCommand.cs
@@ -1,0 +1,12 @@
+﻿using System.Management.Automation;
+
+namespace PSRest.Commands;
+
+[Cmdlet("Reset", "Rest")]
+public sealed class ResetCommand : AnyCmdlet
+{
+    protected override void MyBeginProcessing()
+    {
+        RestEnvironment.Reset();
+    }
+}